change int value to Long for neo4j

+++ nebula-exchange_spark_2.4/src/main/scala/com/vesoft/nebula/exchange/utils/Neo4jUtils.scala
@@ -18,7 +18,7 @@ object Neo4jUtils {
         value.asString()
       }
       case "INTEGER" => {
-        value.asInt().toString
+        value.asLong().toString
       }
       case "FLOAT" | "DOUBLE" => {
         value.asDouble().toString
